requirements.txt
Pillow==6.2.0
future==0.16.0
mercantile==1.1.2